update: 更新文件 audio_control_manager.cpp

+++ services/audio/src/audio_control_manager.cpp
@@ -47,7 +47,7 @@ constexpr int32_t CALL_ENDED_PLAY_TIME = 300;
 constexpr uint64_t UNMUTE_SOUNDTONE_DELAY_TIME = 500000;
 const int16_t MIN_MULITY_ACTIVE_CALL_COUNT = 1;
 const int16_t MIN_DC_MULITY_ACTIVE_CALL_COUNT = 2;
-const int16_t AUDIO_EVENT_MUTED_RINGTONE = 4;
+const int32_t AUDIO_EVENT_MUTED_RINGTONE = 4;
 
 AudioControlManager::AudioControlManager()
     : isLocalRingbackNeeded_(false), ring_(nullptr), tone_(nullptr), sound_(nullptr)
